refactor: move files into components subfolder to be ready for next phase of implementation
