+++ src/tools/icons/career/Education.jsx
