+++ src/tools/icons/career/Internship.jsx
