+++ src/tools/icons/social/LinkedIn.jsx
