+++ src/tools/icons/social/NPM.jsx
