+++ src/tools/icons/technology/AmazonWebServices.jsx
