+++ src/tools/icons/technology/Android.jsx
