+++ src/tools/icons/technology/Angular.jsx
