+++ src/tools/icons/technology/Git.jsx
