+++ src/tools/icons/technology/Redux.jsx
