+++ src/tools/icons/technology/Symfony.jsx
