+++ src/tools/icons/technology/TypeScript.jsx
